[Security Solution] Move rule actions field and its validation into the common folder (#215494)
## Summary

With this PR we make `RuleActionsField` component and relevant
validations reusable outside of and not bound to the rules management.

As part of the Attack Discovery Scheduling
[feature](https://github.com/elastic/security-team/issues/10142) we
would like to be able to setup schedules (similar to detection rules,
just named differently within the feature space) and be able to add
actions (email, slack, webhook etc.).

Currently `RuleActionsField` lives inside the
`x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_creation/components/`
folder. We could just reference it from within the Attack Discovery
folder, but for better code structure it will be good to put it into a
common place.

### Checklist

Check the PR satisfies following conditions.

Reviewers should verify this PR satisfies this list as well.

- [x] [Flaky Test
Runner](https://ci-stats.kibana.dev/trigger_flaky_test_runner/1) was
used on any tests changed
-
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/8075

---------

(cherry picked from commit 760b0cfca324eabaf544f23fdff819bab441576c)

# Conflicts:
#	packages/kbn-babel-preset/styled_components_files.js
