+++ x-pack/solutions/security/plugins/security_solution/public/common/containers/rule_actions/validate_rule_actions_field/utils.ts
