+++ x-pack/solutions/security/plugins/security_solution/public/common/containers/rule_actions/validate_rule_actions_field/validate_rule_actions_field.test.tsx
@@ -8,7 +8,7 @@
 import { validateSingleAction, validateRuleActionsField } from './validate_rule_actions_field';
 import { getActionTypeName, validateMustache, validateActionParams } from './utils';
 import { actionTypeRegistryMock } from '@kbn/triggers-actions-ui-plugin/public/application/action_type_registry.mock';
-import type { FormHook } from '../../../../../shared_imports';
+import type { FormHook } from '../../../../shared_imports';
 jest.mock('./utils');
 
 describe('validate_rule_actions_field', () => {
